Restructured the app.py files by creating separate files to manage different sections of the app
