+++ project/app/__init__.py
@@ -1,0 +1,11 @@
+from flask import Flask
+from flask_migrate import Migrate
+from flask_sqlalchemy import SQLAlchemy
+from config import Config
+
+app = Flask(__name__, static_folder='static', static_url_path='/static',template_folder='templates')
+app.config.from_object(Config)
+db = SQLAlchemy(app)
+migrate = Migrate(app,db)
+
+from app import routes, models
