+++ project/app/static/stylesheets/login-signup.css
@@ -64,3 +64,7 @@ a:hover{
 .btn-secondary{
     background-color: #343a40;
 }
+
+.logo{
+    width: 60%;
+}
