+++ project/app/templates/friends.html
